requirements.txt
requests
PyJWT
python-dotenv
argparse